tests/requirements.txt
asyncio
motor
pytest
pytest-aiohttp
